Revert "170947380 - Send a password reset link to user email"
