+++ package.json
@@ -11,7 +11,6 @@
     "test:watch": "jest --watch",
     "storybook": "start-storybook",
     "coverage": "jest --coverage && cat ./coverage/lcov.info | coveralls",
-    "clear_jest": "jest --clearCache",
     "heroku-postbuild": "webpack -p"
   },
   "repository": {
@@ -44,9 +43,6 @@
     "babel-loader": "^8.0.4",
     "css-loader": "^3.5.2",
     "cssnano": "^4.1.10",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "enzyme-to-json": "^3.4.4",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.1.0",
     "eslint-plugin-import": "^2.20.2",
@@ -60,11 +56,9 @@
     "mini-css-extract-plugin": "^0.9.0",
     "node-sass": "^4.13.1",
     "postcss-loader": "^3.0.0",
-    "react-test-renderer": "^16.13.1",
     "redux-mock-store": "^1.5.4",
     "sass": "^1.26.3",
     "sass-loader": "^8.0.2",
-    "sinon": "^9.0.2",
     "style-loader": "^1.1.4",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.1.1",
@@ -102,7 +96,6 @@
     "redux-logger": "^3.0.6",
     "redux-promise-middleware": "^6.1.2",
     "redux-thunk": "^2.3.0",
-    "remove": "^0.1.5",
     "serve": "^11.3.0",
     "validator": "^13.0.0"
   },
