+++ src/__tests__/login.spec.js
@@ -121,38 +121,38 @@ describe("LOGIN", () => {
     waitForElement(() => fireEvent.click(getByLabelText("Close")));
   });
 
-  // it("should login successfully", async () => {
-  //   const {
-  //     getByLabelText,
-  //     getByText,
-  //     getByTestId,
-  //     container,
-  //     debug,
-  //   } = LoginComponent();
-  //   const email = getByLabelText("email");
-  //   const password = getByLabelText("password");
-  //   const form = container.querySelector("form");
+  it("should login successfully", async () => {
+    const {
+      getByLabelText,
+      getByText,
+      getByTestId,
+      container,
+      debug,
+    } = LoginComponent();
+    const email = getByLabelText("email");
+    const password = getByLabelText("password");
+    const form = container.querySelector("form");
 
-  //   userEvent.type(email, "octopusbn@gmail.com");
-  //   userEvent.type(password, "password");
-  //   form.dispatchEvent(new Event("submit"));
-  // });
-  // it("should change the language to french", async () => {
-  //   const {
-  //     getByLabelText,
-  //     getByText,
-  //     getByTestId,
-  //     container,
-  //     debug,
-  //   } = LoginComponent();
-  //   const french = getByLabelText("french-button");
-  //   fireEvent.click(french);
-  //   waitFor(() =>
-  //     expect(
-  //       getByText(
-  //         "Rendre le voyage et l'hébergement de l'entreprise faciles et pratiques."
-  //       )
-  //     ).toBeTruthy()
-  //   );
-  // });
+    userEvent.type(email, "octopusbn@gmail.com");
+    userEvent.type(password, "password");
+    form.dispatchEvent(new Event("submit"));
+  });
+  it("should change the language to french", async () => {
+    const {
+      getByLabelText,
+      getByText,
+      getByTestId,
+      container,
+      debug,
+    } = LoginComponent();
+    const french = getByLabelText("french-button");
+    fireEvent.click(french);
+    waitFor(() =>
+      expect(
+        getByText(
+          "Rendre le voyage et l'hébergement de l'entreprise faciles et pratiques."
+        )
+      ).toBeTruthy()
+    );
+  });
 });
