+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { ThemeProvider } from '@material-ui/core/styles';
 import App from './App';
-import {theme} from './theme';
+import theme from './theme';
 import './styles/styles.scss';
 import IntlProvider from './languages/components/IntlProvider';
 import store from './redux/store';
