+++ src/languages/messages/en.json
@@ -61,14 +61,5 @@
   "acco_supplier": "Accommodation supplier",
   "travel_admin": "Travel admin",
   "travel_tmember": "Travel team member",
-  "select-role": "You must select a role",
-  
-  "password-gone": "Reset Your Password",
-  "link-promise": "We will send you a password reset link.",
-  "send": "Send",
-  "email": "Email",
-  "networ-error": "Oops! Network Error.",
-  "valid-mail": "Requires valid email",
-  "user-exists-not": "The provided email don't exist, create an account",
-  "process-message": "Please check an email we've sent you to complete the process."
+  "select-role": "You must select a role"
 }
