+++ src/redux/reducers/index.js
@@ -4,8 +4,6 @@ import signUp from "./signupReducer";
 import languageReducer from "./languageReducer";
 import profileReducer from "./profileReducer";
 import roleReducer from "./roleReducer";
-import sendMailReducer from '../reducers/sendMailReducer';
-
 
 const allReducers = combineReducers({
   login: loginReducer,
@@ -13,9 +11,6 @@ const allReducers = combineReducers({
   language: languageReducer,
   profile: profileReducer,
   role: roleReducer,
-  sendEMail: sendMailReducer,
-  
 });
 
 export default allReducers;
-
