--- src/redux/types/PasswordResetTypes.js
@@ -1,11 +1,0 @@
-export const SEND_EMAIL = 'SEND_EMAIL';
-export const SEND_EMAIL_REQUEST = 'SEND_EMAIL_REQUEST';
-export const SEND_EMAIL_SUCCESS = 'SEND_EMAIL_SUCCESS';
-export const SEND_EMAIL_FAILURE = 'SEND_EMAIL_FAILURE';
-
-export const SET_NEW_PASSWORD = 'SET_NEW_PASSWORD';
-export const SET_NEW_PASSWORD_REQUEST = 'SET_NEW_PASSWORD_REQUEST';
-export const SET_NEW_PASSWORD_SUCCESS = 'SET_NEW_PASSWORD_SUCCESS';
-export const SET_NEW_PASSWORD_FAILURE = 'SET_NEW_PASSWORD_FAILURE';
-
-export const CLOSE_MESSAGE = 'CLOSE_MESSAGE';
