+++ src/routes/index.js
@@ -16,8 +16,6 @@ import Requests from "../views/Dashboard/Requests";
 import Profile from "../components/Profile";
 import ProfileSettings from "../components/ProfileSettings";
 import UserRoles from "../views/Dashboard/UserRoles";
-import { SendResetEmailView, PasswordResetProcess } from '../views/SendMailView';
-
 
 export default class index extends Component {
   render() {
@@ -28,9 +26,6 @@ export default class index extends Component {
         <Route path="/verify-email" exact component={Confirm} />
         <Route path="/social-login" exact component={SocialLogin} />
         <Route path="/profile" exact component={Profile} />
-        <Route path="/forgot-password" exact component={SendResetEmailView}/>
-        <Route path="/password-reset-process" exact component={PasswordResetProcess}/>
-
         <DashboardRoot>
           <Route
             component={({ match }) => (
