+++ src/theme.js
@@ -1,7 +1,7 @@
 import { red } from "@material-ui/core/colors";
 import { createMuiTheme } from "@material-ui/core/styles";
 // A custom theme for this app
-export const theme = createMuiTheme({
+const theme = createMuiTheme({
   typography: {
     fontFamily: '"Trebuchet MS", sans-serif',
   },
@@ -17,16 +17,4 @@ export const theme = createMuiTheme({
     },
   },
 });
-
-export const LabelTheme = createMuiTheme({
-	overrides: {
-	  MuiInputLabel: { 
-		root: { 
-		  color: "black",
-		  "&$focused": { // increase the specificity for the pseudo class
-			color: "black"
-		  }
-		}
-	  }
-	}
-  });
+export default theme;
